# [1.0.2] Fix incorrect reported total time of applying a received block or produced block