--- a/libraries/chain/controller.cpp
+++ b/libraries/chain/controller.cpp
@@ -833,7 +833,7 @@ struct pending_state {
    block_stage_type               _block_stage;
    controller::block_status       _block_status = controller::block_status::ephemeral;
    std::optional<block_id_type>   _producer_block_id;
-   controller::block_report       _block_report{};
+   controller::block_report       _block_report;
 
    // Legacy
    pending_state(maybe_session&& s,
@@ -3464,13 +3464,14 @@ struct controller_impl {
          if (s == controller::block_status::incomplete) {
             const auto& id = chain_head.id();
             const auto& new_b = chain_head.block();
-            br.total_time += fc::time_point::now() - start;
+            fc::time_point now = fc::time_point::now();
+            br.total_time += now - start;
 
             ilog("Produced block ${id}... #${n} @ ${t} signed by ${p} "
                  "[trxs: ${count}, lib: ${lib}${confs}, net: ${net}, cpu: ${cpu}, elapsed: ${et}, time: ${tt}]",
                  ("p", new_b->producer)("id", id.str().substr(8, 16))("n", new_b->block_num())("t", new_b->timestamp)
                  ("count", new_b->transactions.size())("lib", chain_head.irreversible_blocknum())("net", br.total_net_usage)
-                 ("cpu", br.total_cpu_usage_us)("et", br.total_elapsed_time)("tt", br.total_time)
+                 ("cpu", br.total_cpu_usage_us)("et", br.total_elapsed_time)("tt", now - br.start_time)
                  ("confs", new_b->is_proper_svnn_block() ? "" : ", confirmed: " + std::to_string(new_b->confirmed)));
          }
 
@@ -3635,7 +3636,7 @@ struct controller_impl {
               ("p", bsp->producer())("id", bsp->id().str().substr(8, 16))("n", bsp->block_num())("t", bsp->timestamp())
               ("count", bsp->block->transactions.size())("lib", bsp->irreversible_blocknum())
               ("net", br.total_net_usage)("cpu", br.total_cpu_usage_us)
-              ("elapsed", br.total_elapsed_time)("time", br.total_time)("latency", (now - bsp->timestamp()).count() / 1000));
+              ("elapsed", br.total_elapsed_time)("time", now - br.start_time)("latency", (now - chain_head.timestamp()).count() / 1000));
          const auto& hb_id = chain_head.id();
          const auto& hb = chain_head.block();
          if (read_mode != db_read_mode::IRREVERSIBLE && hb && hb_id != bsp->id() && hb != nullptr) { // not applied to head
@@ -3659,7 +3660,7 @@ struct controller_impl {
                return false;
             }
 
-            auto start = fc::time_point::now();
+            auto start = fc::time_point::now(); // want to report total time of applying a block
 
             const bool already_valid = bsp->is_valid();
             // Only need to consider voting if not already validated, if already validated then we have already voted.
@@ -3671,6 +3672,9 @@ struct controller_impl {
             const auto& producer_block_id = bsp->id();
 
             start_block( b->timestamp, b->confirmed, new_protocol_feature_activations, s, producer_block_id, fc::time_point::maximum() );
+
+            assert(pending); // created by start_block
+            pending->_block_report.start_time = start;
 
             // validated in create_block_handle()
             std::get<building_block>(pending->_block_stage).trx_mroot_or_receipt_digests() = b->transaction_mroot;
@@ -4459,7 +4463,7 @@ struct controller_impl {
                try {
                   const auto& bsp = *ritr;
 
-                  br = controller::block_report{};
+                  br = controller::block_report();
                   bool applied = apply_block( br, bsp, bsp->is_valid() ? controller::block_status::validated
                                                                        : controller::block_status::complete, trx_lookup );
                   if (!switch_fork && (!applied || check_shutdown())) {
@@ -4494,7 +4498,7 @@ struct controller_impl {
 
                   // re-apply good blocks
                   for( auto ritr = branches.second.rbegin(); ritr != branches.second.rend(); ++ritr ) {
-                     br = controller::block_report{};
+                     br = controller::block_report();
                      apply_block( br, *ritr, controller::block_status::validated /* we previously validated these blocks*/, trx_lookup );
                   }
                   std::rethrow_exception(except);