--- a/libraries/chain/include/eosio/chain/controller.hpp
+++ b/libraries/chain/include/eosio/chain/controller.hpp
@@ -186,7 +186,6 @@ namespace eosio::chain {
             size_t             total_net_usage = 0;
             size_t             total_cpu_usage_us = 0;
             fc::microseconds   total_elapsed_time{};
-            fc::microseconds   total_time{};
             fc::time_point     start_time;
          };