Ported  fix from `main` for incorrect reported total time of applying a received block or produced block.

Only minimum required changes were made, except adding time unit `us` for `elapsed` and `time` for better clarity because `latency` already had `ms` as postfix.

Before the change:
`info  2024-09-30T14:33:33.415 nodeos    controller.cpp:3633           log_applied          ]  Received block bcd746757a3db674... #366563000 @ 2024-08-04T19:14:51.500 signed by eosusacartel [trxs: 0, lib: 366562668, net: 0, cpu: 100, elapsed: 259, time: 869, latency: 4907921915 ms`

After the change:
`info  2024-09-30T14:35:11.723 nodeos    controller.cpp:3627           log_applied          ]  Received block bcd746757a3db674... #366563000 @ 2024-08-04T19:14:51.500 signed by eosusacartel [trxs: 0, lib: 366562668, net: 0, cpu: 100, elapsed: 258 us, time: 554 us, latency: 4908020223 ms]`

Resolves https://github.com/AntelopeIO/spring/issues/828