@ericpassmore 2024-10-01
Note:start
category: Other
component: Internal
summary: Improve reported time for receiving and applying blocks.
Note:end